Implement key derivation using a master password Previously, encryption keys were generated and stored directly in the `credentials/keys/` directory which is a significant security risk.
The application now uses a single master password inputted by the user to access the password manager. The master password uses PBKDF2HMAC with SHA256. This is the standard for password-based encryption.

A cryptographically secure 16-byte salt is also added with the master password during the key derivation function to protect against rainbow table attacks.